api/requirements.txt
# Flask Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Data Processing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dotenv==1.0.0
click==8.1.7
Werkzeug==2.3.7

# Date & Time
python-dateutil==2.8.2

# Validation
marshmallow==3.20.1
email-validator==2.1.0

# JWT Authentication
PyJWT==2.8.0
Flask-JWT-Extended==4.5.3

# Logging
loguru==0.7.2

# Development & Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
flake8==6.1.0
black==23.11.0

# Production Server
gunicorn==21.2.0

# Task Scheduling
APScheduler==3.10.4

# Monitoring (optional)
psutil==5.9.6

# API Documentation (optional)
flasgger==0.9.7.1